requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.3
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kucoin-python==1.0.11
msgpack==1.0.4
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
websockets==10.4
wrapt==1.14.1
zope.interface==5.5.2
